requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-cloud-pubsub==2.19.4
google-cloud-storage==2.14.0
openai==1.23.6
pinecone>=3.0.0
PyMuPDF
